Add files via upload
Project - Daibetes Patients

The report is on the dataset from the National Institute of Diabetes and Digestive and Kidney Diseases. The objective is to diagonstically predict whether a patient has diabetes based on certain diagnostic measurements in the dataset. In particular, all aptients here are females atleast 21 years old of Pima Indian Heritage.

I made the analysis based on the dependent variable(outcome) and the rest of the independent variables.

Looking at the report, the data is collected of 768 patients out of which 268 are diabetes patients. It shows that the major diagnostic measurements which cause diabetes are Glucose and Insulin which are having a higher correlation with the outcome of Diabetes. We can also observe that Skin thickness and DiabetesPedigreeFunction has a negligible relation with the outcome of Diabetes. Also, we can observe that the highest age range that show the sign of Diabetes is 21-30 age group with total of 90.

We can also deduce from the report, the investigation that show the sign of Diabetes.

Skills - Excel and PostgreSQL